refactor: extract local db init routine to new utility handler
